+++ packages/theme/src/index.ts
@@ -1,0 +1,7 @@
+export {
+  applyTheme,
+  type ApplyThemeOptions,
+  type CornerStyle,
+  type ResolvedTheme,
+  type SurfaceMode,
+} from "./applyTheme";